tests/wpt/web-platform-tests/css/css-variables/variable-exponential-blowup.html

<!doctype html>
<title>CSS Variables Test: Exponential blowup doesn't crash</title>
<meta charset="UTF-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://drafts.csswg.org/css-variables/">
<script>
  let css = `
    --v0: "Something really really really long";
  `;
  for (let i = 0; i < 31; ++i)
    css += `--v${i + 1}: var(--v${i}) var(--v${i});`;
  let s = document.createElement("style");
  s.innerHTML = `
    :root { ${css}; }
    :root::before { content: var(--v31); }
  `;
  document.head.appendChild(s);
</script>
PASS if doesn't crash
<script>
  test(function() {
    getComputedStyle(document.documentElement, "::before").content;
    assert_true(true, "Didn't crash");
  });
</script>
